+++ src/surf/network_wifi.cpp
@@ -64,8 +64,8 @@ s4u::Link::SharingPolicy NetworkWifiLink::get_sharing_policy()
 
 void NetworkWifiLink::refresh_decay_bandwidths(){
   // Compute number of STAtion on the Access Point
-  int nSTA=host_rates_.size();
-    
+  int nSTA = static_cast<int>(host_rates_.size());
+
   std::vector<Metric> new_bandwidths;
   for (auto bandwidth : bandwidths_){
     // Instantiate decay model relatively to the actual bandwidth
